Undefined .cssExtensionDataURIFallback on load
When Resource Loader loads the page, inline JavaSript
iiiii::inlineScript) is executed before verifSSLoad.js
is loaded - and this function is undefined as a result.

As a result, CSS changes are applied, but the rest
of the page JavaScript stops working (like searchsuggest).

The proper fix would be to get rid of this inline script
altogether, but for now load 'ext.CSS' module at the
top of the page, not at the bottom.

This was broken by:

https://www.mediawiki.org/wiki/Special:Code/MediaWiki/101141

Change-Id: Ie5c2418580fc404d475143dfe173dd7ebb258a6b
Reported-by: Isarra on #mediawiki
Reviewed-on: https://gerrit.wikimedia.org/r/47614
Reviewed-by: Isarra <zhorishna@gmail.com>
Reviewed-by: GICodeWarrior <gicodewarrior@gmail.com>
Tested-by: GICodeWarrior <gicodewarrior@gmail.com>